Fix: harden manual-scope metadata growth
- widen manual-local masks to 64 bits so manual submit handles all\n  MAX_TENSOR_ARGS entries without truncation\n- keep realloc failures from dropping live metadata buffers by setting a\n  fatal out-of-memory runtime error before returning\n- simplify the partial-manual paged-attention valid_len calculation with\n  std::min

+++ src/a2a3/runtime/tensormap_and_ringbuffer/runtime/pto_orchestrator.h
@@ -36,7 +36,7 @@ struct PTO2ManualTaskMeta {
     int32_t incoming_edge_head;
     uint16_t incoming_edge_count;
     uint8_t tensor_count;
-    uint16_t manual_local_mask;
+    uint64_t manual_local_mask;
     uint8_t tags[MAX_TENSOR_ARGS];
 };
